+++ rosetta/common/errors.go
@@ -2,11 +2,11 @@ package common
 
 import (
 	"fmt"
-	"runtime"
-	"strings"
 
 	"github.com/coinbase/rosetta-sdk-go/types"
+
 	nodeconfig "github.com/harmony-one/harmony/internal/configs/node"
+	"github.com/harmony-one/harmony/internal/utils"
 )
 
 var (
@@ -81,32 +81,6 @@ var (
 	}
 )
 
-// chopPath returns the source filename after the last slash.
-// Inspired by https://github.com/jimlawless/whereami
-func chopPath(original string) string {
-	i := strings.LastIndex(original, "/")
-	if i == -1 {
-		return original
-	}
-	return original[i+1:]
-}
-
-// traceMessage return a string containing the file name, function name
-// and the line number of a specified entry on the call stack.
-// Inspired by https://github.com/jimlawless/whereami
-func traceMessage(depthList ...int) string {
-	var depth int
-	if depthList == nil {
-		depth = 1
-	} else {
-		depth = depthList[0]
-	}
-	function, file, line, _ := runtime.Caller(depth)
-	return fmt.Sprintf("File: %s  Function: %s Line: %d",
-		chopPath(file), runtime.FuncForPC(function).Name(), line,
-	)
-}
-
 // NewError create a new error with a given detail structure
 func NewError(rosettaError types.Error, detailStructure interface{}) *types.Error {
 	newError := rosettaError
@@ -118,7 +92,7 @@ func NewError(rosettaError types.Error, detailStructure interface{}) *types.Erro
 	} else {
 		newError.Details = details
 	}
-	newError.Details["trace"] = traceMessage(2)
+	newError.Details["trace"] = utils.GetCallStackInfo(2)
 	newError.Details["rosetta_version"] = RosettaVersion
 	newError.Details["node_version"] = nodeconfig.GetVersion()
 	return &newError
